Updated minesweeper
I've spent way too much time trying to configure the board and it's cells so I have temporarily put that on hold and will be working on the actual game functions in the mean time. Currently I have a method that sets a specified number of cells to be mines and when you left click on those cells, they highlight red.