Updated minesweeper cells.py
Updated the functionality of left clicking cells. Bombs still don't go off and end the game, however, now clicking on cells that are not bombs show the player how many bombs are surrounding that cells. Not difficult, just takes a cell position and checks all surrounding positions. If cell is in a corner or on the side, there is a single line for loop that takes out it's negative values from the list that we are storing the surrounding cells in.